chore: make sendDistributedTrace fn
